% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregateDepthData.R
\name{AggregateFiles}
\alias{AggregateFiles}
\title{Aggregate Depth Data Files}
\usage{
AggregateFiles(filename_import, filename_export, dir_import, dir_export)
}
\arguments{
\item{filename_import}{Vector of file names.}

\item{filename_export}{File name for combined file.}

\item{dir_import}{Directory for import data.
Default is current working directory.}

\item{dir_export}{Directory for export data.
Default is current working directory.}
}
\value{
Returns a csv into the specified export directory of each file
appended into a single file.
}
\description{
Function for combining high frequency data with depth.
}
\details{
Assumes files have the same column names.
The files are merged with all=TRUE.

One new field (AggFile_Source) is added with the name of the source file.
No other manipulation of the data is performed.

If any columns in the output have the ".x" or ".y" suffix when have duplicate
column names in a file.
}
\examples{
# Data Files
myFile_import <- c("Ellis--1.0m_Water_20180524_20180918.csv"
                  , "Ellis--3.0m_Water_20180524_20180918.csv")
myFile_export <- "Ellis--Combined_Water_20180524_20180918.csv"
myDir_import <- file.path(system.file("extdata", package = "LakeMonitoR"))
myDir_export <- tempdir()
AggregateFiles(filename_import = myFile_import
               , filename_export = myFile_export
               , dir_import = myDir_import
               , dir_export = myDir_export)
}
