% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_depth_means.R
\name{daily_depth_means}
\alias{daily_depth_means}
\title{Depth Profile Daily Means}
\usage{
daily_depth_means(data, col_siteid, col_datetime, col_depth, col_measure)
}
\arguments{
\item{data}{data frame of site id, data/time, depth, and measurement
(e.g., temperature).}

\item{col_siteid}{Column name, SiteID}

\item{col_datetime}{Column name, Date and Time}

\item{col_depth}{Column name, Depth}

\item{col_measure}{Column name, measurement for calculation}
}
\value{
data frame of daily means by depth
}
\description{
Calculates daily means from depth profile data
}
\details{
Interpolates at integer depths.
Daily means with interpolation.

Utilizes the `xts` package.
}
\examples{
# Packages
#library(xts)

# Lake Data
data <- laketemp

# Filter by any QC fields
data <- data[data$FlagV == "P", ]

# name columns
col_siteid   <- "SiteID"
col_datetime <- "Date_Time"
col_depth    <- "Depth_m"
col_measure  <- "Water_Temp_C"

# run function
data_ddm <- daily_depth_means(data
                             , col_siteid
                             , col_datetime
                             , col_depth
                             , col_measure)

# summary
summary(data_ddm)

}
