% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_summary_stats.R
\name{lake_summary_stats}
\alias{lake_summary_stats}
\title{Lake Summary Statistics}
\usage{
lake_summary_stats(data, col_date, col_depth, col_measure, below_threshold = 2)
}
\arguments{
\item{data}{data frame}

\item{col_date}{Column name, Date}

\item{col_depth}{Column name, Depth}

\item{col_measure}{Column name, measurement for calculation}

\item{below_threshold}{Threshold below which to count number of days for
measured value.  Default = 2}
}
\value{
data frame in long format.
}
\description{
Calculates summary statistics on daily depth means.
}
\details{
Input data is assumed to be a single lake location depth profile
with one measurement per day and depth.  The function daily_depth_means can
be used to generate this type of data from continuous data with multiple
measurements per day for each depth.

Calculated statistics are:

* minimum

* maximum

* range

* mean

* median

* quantiles (1, 5, 10, 25, 50, 75, 90, 95, 99)

Statistics are calculated for multiple time periods:

* entire dataset

* year

* month

* year month

A data frame is returned in a hybrid long format.  Statistics will be columns
and time periods will also be a column with the values for each row.  That is,
if time period is month then each month will be a row with each statistic in
a column.
}
\examples{
# data
data <- laketemp_ddm

# Columns
col_date    <- "Date"
col_depth   <- "Depth"
col_measure <- "Measurement"
below_threshold <- 2

# Calculate Stratification
df_lss <- lake_summary_stats(data
                            , col_date
                            , col_depth
                            , col_measure
                            , below_threshold)

# Results
head(df_lss)
}
