% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lakehypso}
\alias{lakehypso}
\title{Lake temperature, hypsography}
\format{
A data frame with 31 observations on the following 2 variables.
 \describe{
   \item{\code{Contour_Depth}}{contour depths (meters)}
   \item{\code{Area}}{a numeric vector of area (square meters)}
 }
}
\usage{
lakehypso
}
\description{
Lake profile depth (m) contour area (m2) for use with rLakeAnalyzer metrics.
}
\keyword{datasets}
