% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LakeMonitoR.R
\docType{package}
\name{LakeMonitoR}
\alias{LakeMonitoR}
\title{Lake Monitor Tools}
\description{
Suite of functions and tools for lake monitoring.
Aids analysis of temperature and dissolved oxygen depth profiles from lakes.
}
