% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{laketemp}
\alias{laketemp}
\title{Lake temperature data}
\format{
\describe{A data frame with 171,414 observations on the following 12 variables.
 \item{\code{DOWLKNUM}}{DOW Lake Number}
 \item{\code{SiteID}}{Site ID}
 \item{\code{Date_Time}}{Date and Time}
 \item{\code{Water_Temp_C}}{water temperature, degree C}
 \item{\code{Depth_m}}{depth, meters}
 \item{\code{Depth_std}}{a numeric vector}
 \item{\code{FlagG}}{QC flag, gross check}
 \item{\code{FlagS}}{QC flag, spike check}
 \item{\code{FlagR}}{QC flag, rate of change check}
 \item{\code{FlagF}}{QC flag, flat line check}
 \item{\code{String}}{a character vector}
 \item{\code{FlagV}}{Overall record level QC flag}
}
}
\usage{
laketemp
}
\description{
Example data containing one year (January 2017 to January 2018) of lake profile
temperature data.  Includes QC flag columns.
}
\keyword{datasets}
