% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{laketemp_ddm}
\alias{laketemp_ddm}
\title{Lake temperature, daily depth means}
\format{
A data frame with 756 observations on the following 3 variables.
 \describe{
   \item{\code{Date}}{a Date, February 2017}
   \item{\code{Depth}}{a numeric vector of depths (meters)}
   \item{\code{Measurement}}{a numeric vector of temperature (degrees C)}
 }
}
\usage{
laketemp_ddm
}
\description{
Lake profile temperature data created from `laketemp` using
`LakeMonitoR::daily_depth_means`.
}
\keyword{datasets}
