% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{laketest}
\alias{laketest}
\title{Test Lake data}
\format{
A data frame with 7,338 observations on the following 5 variables.
 \describe{
   \item{\code{SiteID}}{a character vector}
   \item{\code{Date.Time}}{a character vector}
   \item{\code{DO_conc}}{a numeric vector of dissolved oxygen (mg/L)}
   \item{\code{temp_F}}{a numeric vector of temerature (Fahrenheit)}
   \item{\code{Depth}}{a numeric vector of depth (meters)}
 }
}
\usage{
laketest
}
\description{
Lake profile temperature and dissolved oxygen data.
}
\keyword{datasets}
