% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{laketest_wind}
\alias{laketest_wind}
\title{Test Lake wind data}
\format{
A data frame with 2,446 observations on the following 3 variables.
 \describe{
   \item{\code{SiteID}}{a character vector}
   \item{\code{Date.Time}}{a character vector}
   \item{\code{WSPD}}{a numeric vector of average wind speed (m/s)}
 }
}
\usage{
laketest_wind
}
\description{
Lake wind data for plotting
}
\keyword{datasets}
