% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oxythermal.R
\name{plot_oxythermal}
\alias{plot_oxythermal}
\title{Plot Oxythermal}
\usage{
plot_oxythermal(
  data,
  col_datetime,
  col_depth,
  col_temp,
  col_do,
  thresh_temp = 30,
  operator_temp = ">=",
  thresh_do = 3,
  operator_do = ">=",
  lab_datetime = NA,
  lab_depth = NA,
  lab_temp = NA,
  lab_title = NA,
  boo_interpolate = FALSE,
  boo_contours = FALSE
)
}
\arguments{
\item{data}{data frame of site id (optional), data/time, depth
, and measurement (e.g., temperature).}

\item{col_datetime}{Column name, Date Time}

\item{col_depth}{Column name, Depth}

\item{col_temp}{Column name, Temperature}

\item{col_do}{Column name, Dissolved Oxygen}

\item{thresh_temp}{Threshold for temperature, Default = 30}

\item{operator_temp}{Operator for temperature, Default is >=
Valid values of >=, >, <=, <}

\item{thresh_do}{Threshold for dissolved oxygen, Default is 3}

\item{operator_do}{Operator for dissolved oxygen, Default is <=
Valid values of >=, >, <=, <}

\item{lab_datetime}{Plot label date time (x-axis), Default = col_datetime}

\item{lab_depth}{Plot label for depth (y-axis), Default = col_depth}

\item{lab_temp}{Plot label for temperatue (legend), Default = col_measure}

\item{lab_title}{Plot title, Default = NA}

\item{boo_interpolate}{Boolean as to use interpolate in geom_raster,
Default = FALSE}

\item{boo_contours}{Boolean to draw contours, Default = FALSE}
}
\value{
a ggplot object
}
\description{
Generates a heat map of temperature from depth profile data.
The data is filtered for temperature and dissolved oxygen levels prior to
plotting.
}
\details{
Can be used with any parameters.  A plot is returned that can be
saved with ggsave(filename).

Labels (and title) are function input parameters.  If they are not used
the plot will not be modified.

The default theme is theme_bw().

The plot is created with ggplot2::geom_raster().  Interpolation is possible
with boo_interpolate (default is FALSE).

The returned object is a ggplot object so it can be further manipulated.
}
\examples{
# Data (Test Lake)
data         <- laketest

# Column Names
col_datetime <- "Date.Time"
col_depth    <- "Depth"
col_temp     <- "temp_F"
col_do       <- "DO_conc"

# Data values
thresh_temp <- 68
operator_temp <- ">="
thresh_do <- 3
operator_do <- ">="

# Plot Labels
lab_datetime <- "Date Time"
lab_depth    <- "Depth (m)"
lab_temp     <- "Temp (F)"
lab_title    <- "Test Lake"

# Create Plot
p_ot <- plot_oxythermal(data = data
                        , col_datetime = col_datetime
                        , col_depth = col_depth
                        , col_temp = col_temp
                        , col_do = col_do
                        , thresh_temp = thresh_temp
                        , operator_temp= operator_temp
                        , thresh_do = thresh_do
                        , operator_do = operator_do
                        , lab_datetime = lab_datetime
                        , lab_depth = lab_depth
                        , lab_temp = lab_temp
                        , lab_title = lab_title)

# Print Plot
print(p_ot)

# Demo ability to tweak the plot
## move gridlines on top of plot
p_ot + ggplot2::theme(panel.ontop = TRUE
               , panel.background = element_rect(color = NA, fill = NA))
## Add caption
p_ot + ggplot2::labs(caption = "Gray = Areas outside of given temp and DO
                     values.")

## Add Subtitle
myST <- paste0("Temp ", operator_temp, " ", thresh_temp
             , " and DO ", operator_do, " ", thresh_do)
p_ot + ggplot2::labs(subtitle = myST)

# save plot to temp directory
tempdir() # show the temp directory
ggplot2::ggsave(file.path(tempdir(), "TestLake_plotOxythermal.png"))
}
