% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{run LakeMonitoR Shiny Example}
\usage{
runShiny()
}
\description{
Launches Shiny app for LakeMonitoR
}
\details{
The Shiny app based on the R package LakeMonitoR is included in the R package.
This function launches that app.

The Shiny app is online at:
https://tetratech-wtr-wne.shinyapps.io/LakeMonitoR/
}
\examples{
\dontrun{
# Run Function
runShiny()
}
}
