% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_files_xls.R
\name{save_data_xls}
\alias{save_data_xls}
\title{Save data to multi-worksheet Excel file}
\usage{
save_data_xls(data_names, path, col_names = TRUE, format_headers = TRUE)
}
\arguments{
\item{data_names}{data frame or named list of data frames}

\item{path}{Path, including filename, for Excel file.}

\item{col_names}{Logical to include column names.  Default = TRUE}

\item{format_headers}{Logical to format headers (bold and centered).
Default = TRUE}
}
\value{
a saved Excel file with named data saved to individual worksheets.
}
\description{
Saves multiple data frames to a single Excel file.
}
\details{
Saves multiple data frames to a multiple worksheets in a single Excel
file.

The data must be loaded into R.

The `writexl` package is used so that Java is not necessary.
}
\examples{
# Data (Test Lake)
data         <- laketest


path <- tempfile(fileext = ".xlsx")
data_names <- list


# Save to Excel
save_data_xls(path
              , data_names
              , col_names = TRUE
              , format_headers = TRUE)
}
