% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification.R
\name{stratification}
\alias{stratification}
\title{Lake Stratification}
\usage{
stratification(
  data,
  col_date,
  col_depth,
  col_measure,
  min_days = 1,
  units_depth = "m",
  units_measure = "C"
)
}
\arguments{
\item{data}{data frame}

\item{col_date}{Column name, Date}

\item{col_depth}{Column name, Depth}

\item{col_measure}{Column name, measurement for calculation}

\item{min_days}{Minimum number of consecutive days to be classified as
stratification.  Default = 1.}

\item{units_depth}{Units for depth column (m or ft).  Default = "m".}

\item{units_measure}{Unites for measurement column (C or F).  Default = "C".}
}
\value{
List with 3 elements, 2 data frames and a ggplot
Stratification_Dates includes dates and T/F for Stratified.
Stratification_Events includes a row for each stratified event along with
start and end dates and number of days.
Stratifcation_Plot is a ggplot with lines for each event by date (x)
and year (y).
}
\description{
Calculates stratification on daily depth means.
}
\details{
Uses the output of `daily_depth_means` and assumes units are
degree Celcius and meters.

Calculation is defined as greater than a 1 degree (C) difference over
1 meter anywhere in the water column.

A list is returned with two data frames and a plot.

The first data frame has the dates and TRUE or FALSE for stratification.

The second data frame has the start and end dates, time span, and minimum
value for each stratification event.

The plot is a ggplot object with lines for each stratification event.
The minimum value is shown on the plot for each plot as a red circle.

For the minimum value for each event the value, date, depth, and number of
occurrences appears in the second data frame.  Only the first value is shown.

Units should be metric (meters and celsius).  If use Imperial units (feet and
fahrenheit) they will be converted.  Accepted abberviations are m, ft, C, and
F.

Input data is assumed to be a single lake location depth profile.
}
\examples{
# data
data <- laketemp_ddm

# Columns
#col_siteid  <- "SiteID"
col_date    <- "Date"
col_depth   <- "Depth"
col_measure <- "Measurement"

# Calculate Stratification
ls_strat <- stratification(data
                           , col_date
                           , col_depth
                           , col_measure
                           , min_days = 1 )

# Results, Stratification Dates
head(ls_strat$Stratification_Dates)

# Results, Stratification Events
ls_strat$Stratification_Events

# Results, Stratification Plot
p <- ls_strat$Stratification_Plot
print(p)
}
