% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification.R
\name{stratification}
\alias{stratification}
\title{Lake Stratification}
\usage{
stratification(data, col_date, col_depth, col_measure, min_days = 20)
}
\arguments{
\item{data}{data frame}

\item{col_date}{Column name, Date}

\item{col_depth}{Column name, Depth}

\item{col_measure}{Column name, measurement for calculation}

\item{min_days}{Minimum number of consecutive days to be classified as
stratification.  Default = 20.}
}
\value{
List of 2 data frames.
Stratification_Dates includes dates and T/F for Stratified.
Stratification_Events includes a row for each stratified event along with
start and end dates and number of days.
}
\description{
Calculates stratification on daily depth means.
}
\details{
Uses the output of `daily_depth_means` and assumes units are
degree Celcius and meters.

Calculation is defined as greater than a 1 degree (C) difference over
1 meter anywhere in the water column.

A list is returned with two data frames.  The first has the dates and
TRUE or FALSE for stratification  The second has the start and end dates and
time span for each stratification event.

Input data is assumed to be a single lake location depth profile.
}
\examples{
# data
data <- laketemp_ddm

# Columns
#col_siteid  <- "SiteID"
col_date    <- "Date"
col_depth   <- "Depth"
col_measure <- "Measurement"

# Calculate Stratification
ls_strat <- stratification(data, col_date, col_depth, col_measure, min_days = 20 )

# Results, Stratification Dates
head(ls_strat$Stratification_Dates)

# Results, Stratification Events
ls_strat$Stratification_Events

}
